requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2.11
django-heroku==0.3.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
gunicorn==20.1.0
psycopg2==2.9.3
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
typing-extensions==4.1.1
whitenoise==6.0.0
